Add create Twitter notificator modal box in Gamecraft UI

+++ gamecraft-ui/src/main/resources/static/js/ui_operations.js
@@ -278,14 +278,27 @@ function fillTwitterBotsTable() {
     var bots = getTwitterBots();
     $(function() {
         $.each(bots, function(i, item) {
-            var tr = $('<tr>').append(
-                $('<tr>'),
-                $('<td>').text(item.id),
-                $('<td>').text(item.twitterBotName),
-                $('<td>').text(item.twitterBotDescription),
-                $('<td>').html("<button type=\"button\" class=\"btn btn-danger btn-xs\" onclick=\"deleteTwitterBot('" +item.id + "'); location.reload(); \"><span class=\"glyphicon glyphicon-remove\"></span> </button>")
-            );
-            $(".table").append(tr.html());
+            if (isAdmin(getUsername())) {
+
+                var tr = $('<tr>').append(
+                    $('<tr>'),
+                    $('<td>').text(item.id),
+                    $('<td>').text(item.twitterBotName),
+                    $('<td>').text(item.twitterBotDescription),
+                    $('<td>').html("<button type=\"button\" class=\"btn btn-danger btn-xs\" onclick=\"deleteTwitterBot('" + item.id + "'); location.reload(); \"><span class=\"glyphicon glyphicon-remove\"></span> </button>")
+                );
+                $(".table").append(tr.html());
+            }
+            else {
+                var tr = $('<tr>').append(
+                    $('<tr>'),
+                    $('<td>').text(item.id),
+                    $('<td>').text(item.twitterBotName),
+                    $('<td>').text(item.twitterBotDescription),
+                    $('<td>').html("")
+                );
+                $(".table").append(tr.html());
+            }
         });
     });
 }
